requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
